+++ backend/routers/gardensRouter.js
@@ -1,11 +1,18 @@
 const express = require('express');
 const { getAllGardens, getGardensForAuthorizedUser } = require('../controllers/gardensController');
 
+const authMiddleware = require('../authMiddleware');
+
 const router = express.Router();
 
 // accepts query parameter gardenId
-router.get('/all', getAllGardens);
+// TODO: ADD GARDEN OWNER NAMES
+router.get('/all', authMiddleware, getAllGardens);
 
-router.get('/', getGardensForAuthorizedUser);
+router.get('/', authMiddleware, getGardensForAuthorizedUser);
+
+// TODO: GET Garden members and their roles by garden
+
+// TODO: GET garden by authorized user
 
 module.exports = router;
