+++ backend/routers/updatesRouter.js
@@ -1,8 +1,10 @@
 const express = require('express');
 const { getUpdatesForAuthenticatedUser } = require('../controllers/updatesController');
 
+const authMiddleware = require('../authMiddleware');
+
 const router = express.Router();
 
-router.get('/', getUpdatesForAuthenticatedUser);
+router.get('/', authMiddleware, getUpdatesForAuthenticatedUser);
 
 module.exports = router;
