+++ frontend/app/src/main/java/com/plotpals/client/CreateProfileActivity.java
@@ -133,7 +133,7 @@ public class CreateProfileActivity extends AppCompatActivity {
      */
     private void skipProfileCreationIfUserProfileAlreadyExists() {
         RequestQueue volleyQueue = Volley.newRequestQueue(this);
-        String url = "http://10.0.2.2:8081/profiles?profileId=" + googleProfileInformation.getAccountUserId();
+        String url = "http://10.0.2.2:8081/profiles/all?profileId=" + googleProfileInformation.getAccountUserId();
 
         Request<?> jsonObjectRequest = new JsonObjectRequest(
                 Request.Method.GET,
