+++ frontend/app/src/main/java/com/plotpals/client/GardenSearchActivity.java
@@ -101,7 +101,7 @@ public class GardenSearchActivity extends AppCompatActivity {
 
     private void requestGardens() {
         RequestQueue volleyQueue = Volley.newRequestQueue(this);
-        String url = "http://10.0.2.2:8081/gardens/";
+        String url = "http://10.0.2.2:8081/gardens/all";
 
         Request<?> jsonObjectRequest = new JsonObjectRequest(
             Request.Method.GET,
