+++ frontend/app/src/main/java/com/plotpals/client/MapsActivity.java
@@ -374,7 +374,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
 
     private void requestGardens() {
         RequestQueue volleyQueue = Volley.newRequestQueue(this);
-        String url = "http://10.0.2.2:8081/gardens/";
+        String url = "http://10.0.2.2:8081/gardens/all";
 
         Request<?> jsonObjectRequest = new JsonObjectRequest(
             Request.Method.GET,
